[ADD] auth_oauth: added Microsoft Azure auth provider This commit enables user to login with their Microsoft Account.
We have used implicit type authentication for all other providers,
which directly gives access_token for fetching user details.

However, to implement implicit type auth in case of Microsoft, an
extra parameter 'nonce' needs to be passed in the request to fetch
the access_token and id_token (id_token is used for fetching basic
user details). Here 'nonce', should be a randomized value which can
be verified in response to mitigate token replay attacks, and thus
have to be stored somewhere if we want to verify it.

Also, instead of using implicit type auth, it is recommended for
web-apps to use Authorization Code Grant flow for authentication,
which is more secure than the former one.

For both above reasons, this commit makes it possible to use the
Authorization Code Grant flow for all providers, and uses it
by default for Microsoft Azure.